+++ backend/src/main/java/com/tripmarket/domain/member/dto/MemberResponseDTO.java
@@ -2,25 +2,20 @@ package com.tripmarket.domain.member.dto;
 
 import com.tripmarket.domain.member.entity.Member;
 
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class MemberResponseDTO {
-	private Long id;
-	private String email;
-	private String name;
-	private String imageUrl;
-	private Boolean hasGuideProfile;
-
+public record MemberResponseDTO(
+	Long id,
+	String email,
+	String name,
+	String imageUrl,
+	Boolean hasGuideProfile
+) {
 	public static MemberResponseDTO from(Member member) {
-		return MemberResponseDTO.builder()
-			.id(member.getId())
-			.email(member.getEmail())
-			.name(member.getName())
-			.imageUrl(member.getImageUrl())
-			.hasGuideProfile(member.getHasGuideProfile())
-			.build();
+		return new MemberResponseDTO(
+			member.getId(),
+			member.getEmail(),
+			member.getName(),
+			member.getImageUrl(),
+			member.getHasGuideProfile()
+		);
 	}
 }
